fix(pi-telemetry): break index.ts import cycle so the extension loads
index.ts re-exports the extension factory (export { default } from
'./extension.js') while extension.ts, otel.ts, and langfuse/config.ts imported
NoopRuntimeEventExporter as a runtime value back from index.ts. That makes
index.ts a circular-import target: a static/bundling loader (Pi's extension
loader, and jiti) evaluates the entry before the cycle resolves and reads the
re-exported default as undefined, so the extension fails to load with
"undefined is not an object (evaluating '_extension.default')". Node's native
ESM tolerates the cycle via live bindings, which is why vitest's native loader
and the package's own tests never caught it.

Move NoopRuntimeEventExporter and CompositeRuntimeEventExporter into a leaf
module (src/exporters.ts) and re-export them from index.ts, so nothing imports
runtime values back from the entry. The public @amaster.ai/pi-telemetry entry
point still exports both classes, so the documented contract is unchanged.

Regression test: un-skip pi-telemetry in tests/extension-e2e.test.ts, whose
jiti static loader reproduces the undefined-default failure and now passes.
The stale comment blaming "jiti 2.7 static interop" is corrected to name the
circular import (pi-task-scheduler has the same shape and stays skipped).
